Fixed another crash. OnHandleIntent was called with null intent. This can happen when START_STICKY is returned in OnStartCommand. Solution is to return START_STICKY_COMPATIBILITY instead.
Another crash happened in SendAndScheduleNextKA because m_tester was null. This can happen when the service restarts, and start ka testing intent hasn't been sent through this restarted service. Now, AlarmManager continues to send the KA intents, so we may end up in a crash. Fixed the crash now. (Note that the right solution would be to ensure that the KA test intent is also triggered when the service restarts. This chould be done as follows: when NULL intent is received (revert to START_STICKY), change to START_KA_TESTING intent)